move commands out of subfolder
