+++ lib/commands/twitter.js
@@ -1,7 +1,7 @@
-const { getTwitterList } = require('../../modules/twitter.js');
+const { getTwitterList } = require('../modules/twitter.js');
 const commando = require('discord.js-commando');
-const constants = require('../../constants.js');
-const util = require("../../util.js");
+const constants = require('../constants.js');
+const util = require("../util.js");
 
 const MAX_QUEUE_SIZE = 20
 const streamOptions = { seek: 0, volume: 1 };
