+++ lib/commands/youtube.js
@@ -1,9 +1,9 @@
-const { youtube_id, youtube, getVideoPromise } = require('../../modules/youtube.js');
+const { youtube_id, youtube, getVideoPromise } = require('../modules/youtube.js');
 const commando = require('discord.js-commando');
-const constants = require('../../constants.js');
+const constants = require('../constants.js');
 const ytdl = require('ytdl-core');
 const Request = require('request');
-const util = require("../../util.js");
+const util = require("../util.js");
 
 const MAX_QUEUE_SIZE = 20
 const streamOptions = { seek: 0, volume: 1 };
