+++ lib/index.js
@@ -54,10 +54,6 @@ client.login(process.env.TOKEN)
         getSocialMedia(pg_client);
       })
     }
-
-    let msg = `I'm updated! v${require('../package.json').version}`;
-
-    client.channels.get((process.env.npm_config_test !== undefined) ? TEST_ID : REOL_CHAN_ID).send(msg);
   }, (reason) => { catchErrors(reason) })
 
 var catchErrors = (reason) => {
